move the Wallet Configuration Panel into its own component, and use it within Greeter
